% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geer.R
\docType{data}
\name{aids}
\alias{aids}
\title{Aids Data}
\format{A data frame with 2376 rows and 8 variables}
\usage{
data(aids)
}
\description{
The aids dataset comprises a total of 2376 CD4+ cell counts for 369 HIV
infected men with a follow up period of approximately eight and half year.
The number of measurements for each individual varies from 1 to 12 and the
times are not equally spaced. The CD4+ cell data are highly unbalanced.
}
\details{
\itemize{
  \item id: subject id
  \item time: measurement time
  \item cd4: CD4+ cell count
}
}
\keyword{datasets}
