% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{$,gee_jmcm-method}
\alias{$,gee_jmcm-method}
\title{Extract parts of gee_jmcm.}
\usage{
\S4method{$}{gee_jmcm}(x, name)
}
\arguments{
\item{x}{an object of gee_jmcm class}

\item{name}{member function of gee_jmcm class}
}
\description{
Extract parts of gee_jmcm.
}
