% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{$,ipw-method}
\alias{$,ipw-method}
\title{Extract parts of ipw.}
\usage{
\S4method{$}{ipw}(x, name)
}
\arguments{
\item{x}{an object of ipw class}

\item{name}{member function of ipw class}
}
\description{
Extract parts of ipw.
}
