% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geer.R
\docType{data}
\name{cattle}
\alias{cattle}
\title{Cattle Data}
\format{A data frame with 660 rows and 4 variables}
\usage{
data(cattle)
}
\description{
Kenward (1987) reported an experiment in which cattle were assigned randomly
to two treatment groups A and B, and their body weights were recorded in
kilogram. Thirty animals received treatment A and another 30 received
treatment B. The animals were weighted 11 times over a 133-day period; the
first 10 measurements for each animal were made at two-week intervals and the
last measurement was made one week later. Since no observation was missing,
it is considered to be a balanced longitudinal dataset.
}
\details{
\itemize{
  \item id: subject id
  \item day: measurement time
  \item group: Treatment A or Treatment B
  \item weight: cattle weight
}
}
\keyword{datasets}
