% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{gee_jmcm-class}
\alias{gee_jmcm-class}
\title{Class gee_jmcm}
\description{
Class \code{gee_jmcm} defines a Modified Cholesky decomposition (MCD) based
joint mean covariance model within the framework of (weighted) generalised
estimating equations.
}
