% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{geerMod-class}
\alias{geerMod-class}
\title{Class "geerMod" of Fitted GEE-MCD/WGEE-MCD  Models.}
\description{
Class "geerMod" of Fitted GEE-MCD/WGEE-MCD  Models.
}
\section{Slots}{

\describe{
\item{\code{call}}{the matched call}

\item{\code{opt}}{the optimization result returned by optimizeGeer}

\item{\code{args}}{arguments m, Y, X, Z, W, H, time}

\item{\code{triple}}{an integer vector of length three containing the degrees of the
three polynomial functions for the mean structure, the log innovation
-variances and the autoregressive coefficients.}

\item{\code{rho}}{a parameter used in the 'working' covariance structure.}

\item{\code{devcomp}}{the deviance components list.}
}}

