% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GenerateCattleMAR}
\alias{GenerateCattleMAR}
\title{Generate Cattle B Data with Pre-specified Dropout Rate}
\usage{
GenerateCattleMAR(dropout.rate)
}
\arguments{
\item{dropout.rate}{the dropout rate at the fifth measurement.}
}
\description{
We assume that all cattle have the first four observations and
  there is a certain chance that some cows will quit the study (dropout) at
  the fifth measurement if their weights are below a certain threshold. The
  threshold value for the weight is chosen so that some fixed rates of MAR
  dropout at the fifth measurement time.
}
