% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ipw_estimation}
\alias{ipw_estimation}
\title{Fit Inverse Probability Weights Model}
\usage{
ipw_estimation(m, Y, order, trace = FALSE)
}
\arguments{
\item{m}{an integer vector of number of measurements for each subject.}

\item{Y}{a vector of responses for all subjects.}

\item{order}{the order for MAR remaining model.}

\item{trace}{whether or not the optimization iteration should be printed.}
}
\description{
Fit inverse probability weights model.
}
