% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geer.R
\docType{methods}
\name{show,geerMod-method}
\alias{show,geerMod-method}
\title{Print information for geerMod-class}
\usage{
\S4method{show}{geerMod}(object)
}
\arguments{
\item{object}{a fitted GEE-MCD/WGEE-MCD model of class "geerMod", i.e.,
typically the result of geer().}
}
\description{
Print information for geerMod-class
}
