% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geer.R
\name{geer}
\alias{geer}
\title{Fit Joint Mean-Covariance Models}
\usage{
geer(formula, data = NULL, triple = c(3, 3, 3), method = c("gee-mcd",
  "wgee-mcd"), corr.struct = c("id", "cs", "ar1"), rho = 0.5,
  ipw.order = 1, control = geerControl(), start = NULL)
}
\arguments{
\item{formula}{a two-sided linear formula object describing the covariates
for both the mean and covariance matrix part of the model, with the response,
the corresponding subject id and measurement time on the left of a operator~,
divided by vertical bars ("|").}

\item{data}{a data frame containing the variables named in formula.}

\item{triple}{an integer vector of length three containing the degrees of the
three polynomial functions for the mean structure, the log innovation
-variances and the autoregressive or moving average coefficients when 'mcd'
or 'acd' is specified for cov.method. It refers to the degree for the mean
structure, variances and angles when 'hpc' is specified for cov.method.}

\item{corr.struct}{covariance structure modelling method,
choose 'mcd' (Pourahmadi 1999), 'acd' (Chen and Dunson 2013, not available yet)
or 'hpc' (Zhang et al. 2015, not available yet).}

\item{rho}{a parameter used in the sandwich 'working' covariance structure.}

\item{control}{a list (of correct class, resulting from geerControl())
containing control parameters, see the *geerControl documentation for
details.}

\item{start}{starting values for the parameters in the model.}
}
\description{
Fit a joint mean-covariance model to longitudinal data within
the within the framework of generalised estimating equations (GEE), via the
Newton-Raphson method.
}
\examples{
cattleA <- cattle[cattle$group=='A', ]
fit.mcd <- geer(weight | id | I(ceiling(day/14 + 1)) ~ 1 | 1,
data=cattleA, triple = c(8, 2, 2), rho = 0.5, corr.struct = 'cs')
}

